Moves request interface extension methods into the same file
